+++ test/transformRecipients.js
@@ -61,13 +61,39 @@ describe('index.js', function() {
           });
       });
 
-    it('should transform recipients according a domain wildcard mapping',
+    it('should transform recipients according to a domain wildcard mapping',
       function(done) {
         var data = {
           recipients: ["info@EXAMPLE.com"],
           config: {
             forwardMapping: {
               "@example.com": [
+                "jim@example.com",
+                "jane@example.com"
+              ]
+            },
+            log: console.log
+          }
+        };
+        index.transformRecipients(data)
+          .then(function(data) {
+            assert.equal(data.recipients[0],
+              "jim@example.com",
+              "parseEvent made 1/2 substitutions");
+            assert.equal(data.recipients[1],
+              "jane@example.com",
+              "parseEvent made 2/2 substitutions");
+            done();
+          });
+      });
+
+    it('should transform recipients according to a user wildcard mapping',
+      function(done) {
+        var data = {
+          recipients: ["info@foo.com"],
+          config: {
+            forwardMapping: {
+              info: [
                 "jim@example.com",
                 "jane@example.com"
               ]
@@ -104,6 +130,28 @@ describe('index.js', function() {
           log: console.log
         };
         index.transformRecipients(data);
+      });
+
+    it('should support matching a name without domain',
+      function(done) {
+        var data = {
+          recipients: ["info"],
+          config: {
+            forwardMapping: {
+              info: [
+                "jim@example.com"
+              ]
+            }
+          },
+          log: console.log
+        };
+        index.transformRecipients(data)
+          .then(function(data) {
+            assert.equal(data.recipients[0],
+              "jim@example.com",
+              "parseEvent made substitution");
+            done();
+          });
       });
 
     it('should exit if the recipient is malformed',